+++ app/src/main/java/au/com/guidebee/morsetoolkit/activity/mario/actors/items/CollectibleItem.java
@@ -19,7 +19,7 @@ public abstract class CollectibleItem extends Sprite implements Collectible {
     }
 
     @Override
-    public boolean overlaps(int x, int y, int width, int height) {
+    public boolean overlaps(float x, float y, int width, int height) {
         return active
                 && x < getX() + getWidth() && x + width > getX()
                 && y < getY() + getHeight() && y + height > getY();
